Import internal library
import format command line as internal library
