+++ src/bioseq_kmers/cli/kmer_analysis.py
@@ -67,11 +67,7 @@ import os
 import time
 import datetime
 import sys
-# For shell command string manipulation
-import shlex
 
-from tqdm import tqdm
-from collections import Counter # count (cf k-mer count)
 # Coloring warning text
 from colorama import init, Fore
 # To reset color
@@ -84,6 +80,7 @@ init(autoreset=True)
 import bioseq_kmers.sequences as seq
 import bioseq_kmers.kmer_stats as kmers
 import bioseq_kmers.background_models as bg
+import bioseq_kmers.utils as utils
 
 ################################################################
 ## CONSTANTS
@@ -103,7 +100,6 @@ _DICT_DEPENDENCIES = {"obs_freq" : ["occ"],
 #####################################
 #   Function: Parse return option   #
 #####################################
-
 
 # User parameter for --return
 def parse_return_option(return_str):
@@ -138,7 +134,6 @@ def parse_return_option(return_str):
 ##############################
 #   Function: Dependencies   #
 ##############################
-
 
 # Automatic addition of dependencies
 def resolve_dependencies(fields):
@@ -180,53 +175,6 @@ def resolve_dependencies(fields):
                     change = True
 
     return list(resolved)
-
-
-#################################################
-#   Function: Defined the format command line   #
-#################################################
-def format_command_line(argv):
-    """
-     Format command line by changing the absolute path to a relative path.
-
-     Args:
-         argv (list): Command-line arguments.
-
-     Returns:
-         str: Reconstructed command line.
-     """
-    # Retrieves the folder from which the script is executed
-    cwd = os.getcwd()
-    # Contains the rebuilt command
-    list_cleaned_command = []
-
-    # The next argument is a file path
-    skip_next = False
-
-    # Iterates through each element of the order
-    for arg in argv:
-
-        # If argument associated with -i or -o
-        if skip_next:
-            # Converts absolute path to relative path
-            rel = os.path.relpath(arg, cwd)
-            list_cleaned_command.append(shlex.quote(rel))
-            # Returns to the initial state
-            skip_next = False
-
-        elif arg in ["-i", "--input", "-o", "--output"]:
-            # Keep the current flag
-            list_cleaned_command.append(arg)
-            # The next argument is a path
-            skip_next = True
-
-        else:
-            # Normal argument
-            list_cleaned_command.append(shlex.quote(arg))
-
-    return " ".join(list_cleaned_command)
-
-
 
 #################
 #   Main code   #
@@ -439,7 +387,7 @@ def main():
 
         ## Parameter
         # Command line
-        command_line = format_command_line(sys.argv)
+        command_line = utils.format_command_line(sys.argv)
         # Write command line
         tsv_file.write(f"; Command\t{command_line}\n;\n")
         # URL in input
