+++ src/bioseq_kmers/cli/markov_from_kmers.py
@@ -19,55 +19,11 @@ from colorama import init, Fore
 ############################
 
 import bioseq_kmers.background_models as markov_bg
+import bioseq_kmers.utils as utils
 
 ################################################################
 ## FUNCTIONS
 ################################################################
-
-
-#################################################
-#   Function: Defined the format command line   #
-#################################################
-def format_command_line(argv):
-    """
-     Format command line by changing the absolute path to a relative path.
-
-     Args:
-         argv (list): Command-line arguments.
-
-     Returns:
-         str: Reconstructed command line.
-     """
-    # Retrieves the folder from which the script is executed
-    cwd = os.getcwd()
-    # Contains the rebuilt command
-    list_cleaned_command = []
-
-    # The next argument is a file path
-    skip_next = False
-
-    # Iterates through each element of the order
-    for arg in argv:
-
-        # If argument associated with -i or -o
-        if skip_next:
-            # Converts absolute path to relative path
-            rel = os.path.relpath(arg, cwd)
-            list_cleaned_command.append(shlex.quote(rel))
-            # Returns to the initial state
-            skip_next = False
-
-        elif arg in ["-i", "--input", "-o", "--output"]:
-            # Keep the current flag
-            list_cleaned_command.append(arg)
-            # The next argument is a path
-            skip_next = True
-
-        else:
-            # Normal argument
-            list_cleaned_command.append(shlex.quote(arg))
-
-    return " ".join(list_cleaned_command)
 
 ###########################
 #   Function: load kmer   #
@@ -149,7 +105,7 @@ def main():
     with open(output_path, "w") as tsv_file:
         ## Parameter
         # Command line
-        command_line = format_command_line(sys.argv)
+        command_line = utils.format_command_line(sys.argv)
         # Write command line
         tsv_file.write(f"; markov-from-kmers\t{command_line}\n;\n")
         # URL in input
