+++ src/bioseq_kmers/cli/markov_model.py
@@ -61,12 +61,10 @@ import argparse
 import time
 import datetime
 import os
-# For shell command string manipulation
-import shlex
 import sys
 
 # Coloring warning text
-from colorama import init, Fore
+from colorama import Fore
 
 ############################
 #   Internal libraries     #
@@ -76,54 +74,8 @@ from colorama import init, Fore
 import bioseq_kmers.sequences as seq
 # Markov model background
 import bioseq_kmers.background_models as markov_bg
-
-################################################################
-## FUNCTIONS
-################################################################
-
-#################################################
-#   Function: Defined the format command line   #
-#################################################
-def format_command_line(argv):
-    """
-     Format command line by changing the absolute path to a relative path.
-
-     Args:
-         argv (list): Command-line arguments.
-
-     Returns:
-         str: Reconstructed command line.
-     """
-    # Retrieves the folder from which the script is executed
-    cwd = os.getcwd()
-    # Contains the rebuilt command
-    list_cleaned_command = []
-
-    # The next argument is a file path
-    skip_next = False
-
-    # Iterates through each element of the order
-    for arg in argv:
-
-        # If argument associated with -i or -o
-        if skip_next:
-            # Converts absolute path to relative path
-            rel = os.path.relpath(arg, cwd)
-            list_cleaned_command.append(shlex.quote(rel))
-            # Returns to the initial state
-            skip_next = False
-
-        elif arg in ["-i", "--input", "-o", "--output"]:
-            # Keep the current flag
-            list_cleaned_command.append(arg)
-            # The next argument is a path
-            skip_next = True
-
-        else:
-            # Normal argument
-            list_cleaned_command.append(shlex.quote(arg))
-
-    return " ".join(list_cleaned_command)
+# Format command line
+import bioseq_kmers.utils as utils
 
 
 #################
@@ -241,7 +193,7 @@ def main():
     with open(output_path, "w") as tsv_file:
         ## Parameter
         # Command line
-        command_line = format_command_line(sys.argv)
+        command_line = utils.format_command_line(sys.argv)
         # Write command line
         tsv_file.write(f"; markov-from-seq\t{command_line}\n;\n")
         # URL in input
